[PATCH] [Bug] update makeTransformation to consider svgs where width
I was using the commandline tool to convert svg icons to SF Symbols. There were a few examples where the height of the icon is less than its width, causing the computation of the scale in `Renderer.SFSymbol.swift` `makeTransformation` to be too large. This resulted in the width of the symbol being incorrect.

I have included one of the icon svg tiles (dash.svg) and a test to make sure the bounds are computed properly. I’m not sure if this is the proper fix, but a repro step is very easy. Simply run:

```
SwiftDraw dash.svg —format sfsymbol
```